Phase 2 - Progress 24/10/2021 Part 2
Details:
 - Fix bug when initially approving candidateDetailChange.
 - Fix filter conditions for candidate lists on job portal page (everything required except achievements & interests)
 - Hiring Partners can now archive (save) candidates before contacting them.
 - Add to my list button on job portal candidate list cards dynamic (wordings changed to Added if candidate is already in list).

+++ app/Http/Controllers/Admin/CandidateController.php
@@ -229,7 +229,8 @@ class CandidateController extends Controller
 
         // If candidateDetailChange have cv_file, unlink previous and include in update data.
         if ($candidateDetailChange->hasAttribute('cv_file') && !is_null($candidateDetailChange->cv_file)) {
-            unlink($candidateDetailChange->candidateDetail->cv_file);
+            if ($candidateDetailChange->candidateDetail->hasAttribute('cv_file') && !is_null($candidateDetailChange->candidateDetail->cv_file))
+                unlink($candidateDetailChange->candidateDetail->cv_file);
             $data['cv_file'] = $candidateDetailChange['cv_file'];
         }
